+++ src/neo_grammar_graph/__init__.py
@@ -1,3 +1,3 @@
-__version__ = "0.0.3"
+__version__ = "0.0.4"
 
 from neo_grammar_graph.gg import NeoGrammarGraph  # noqa