+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 
 ## [unreleased]
 
+## [0.0.4] - 2023-04-19
+
+### Added
+
+- Computation of all paths between two grammar symbols.
+
+## [0.0.2] - 2023-04-19
+
+### Added
+
+- Shortest path computation.
+
+## [0.0.1] - 2023-04-18
+
+### Added
 ## [0.0.3] - 2023-04-19
 
 ### Added